requirements.txt
pandas==1.4.4
pytest==7.1.2
torchvision~=0.15.0.dev20221101

imageio~=2.19.3
matplotlib~=3.5.2
numpy~=1.23.1
pip~=22.2.2
zlib~=1.2.12
wheel~=0.37.1
brotli~=1.0.9
openssl~=1.1.1q
cryptography~=37.0.1
py~=1.11.0
tornado~=6.2
typing_extensions~=4.3.0
setuptools~=63.4.1
scipy~=1.7.3
tqdm~=4.64.1
dill~=0.3.4
yaml~=0.2.5
pyyaml~=6.0